made a generic Model class with logistic regression and NN as child classes, modified  classification pipeline to be generic
